final/policy change/regulatory-sentinel/requirements.txt
# Core (requires Python 3.11+)
langgraph>=0.0.40
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0

# Database
sqlalchemy>=2.0.0
chromadb>=0.4.0

# PDF Processing
pymupdf>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
